+++ src/rules/sort-object-properties-by-type.ts
